requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
requests==2.32.3
groq==0.18.0
Jinja2==3.1.5
python-multipart==0.0.9
langchain-core==0.3.35
langchain-groq==0.2.4
python-dotenv==1.0.1
python-dotenv==1.0.1